third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-await/async-callstack-async-await-expected.txt

Tests asynchronous call stacks for async functions.

Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) doTestChainedPromises (async-callstack-async-await.html:100)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:99)
    3) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) doTestChainedPromises (async-callstack-async-await.html:102)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:101)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:99)
    3) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)

Call stack:
    0) doTestChainedPromises (async-callstack-async-await.html:104)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await.html:103)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:101)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:99)
    3) testFunctionTimeout (async-callstack-async-await.html:50)

Call stack:
    0) doTestChainedPromises (async-callstack-async-await.html:106)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await.html:105)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await.html:103)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:101)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)

Call stack:
    0) errorCallback (async-callstack-async-await.html:60)
    1) doTestPromiseReject (async-callstack-async-await.html:191)
    [Promise.reject]
    0) doTestPromiseReject (async-callstack-async-await.html:188)
    1) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) errorCallback (async-callstack-async-await.html:60)
    1) doTestRejectFromChain (async-callstack-async-await.html:171)
    [Promise.reject]
    0) rejectPromise (async-callstack-async-await.html:16)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:19)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) rejectFromChain (async-callstack-async-await.html:160)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) rejectFromChain (async-callstack-async-await.html:159)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)

Call stack:
    0) errorCallback (async-callstack-async-await.html:60)
    1) doTestSettledPromisesRejected (async-callstack-async-await.html:92)
    [Promise.reject]
    0) doTestSettledPromisesRejected (async-callstack-async-await.html:89)
    1) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) errorCallback (async-callstack-async-await.html:60)
    1) doTestThrowFromChain (async-callstack-async-await.html:151)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) throwFromChain (async-callstack-async-await.html:138)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) throwFromChain (async-callstack-async-await.html:137)
    3) doTestThrowFromChain (async-callstack-async-await.html:148)
    4) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)

Call stack:
    0) promiseCallback (async-callstack-async-await.html:68)
    1) doTestPromiseConstructor (async-callstack-async-await.html:66)
    2) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) thenCallback (async-callstack-async-await.html:55)
    1) doTestChainedPromises (async-callstack-async-await.html:107)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromises (async-callstack-async-await.html:107)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await.html:105)
    [Promise.resolve]
    0) doTestChainedPromises (async-callstack-async-await.html:103)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)

Call stack:
    0) thenCallback (async-callstack-async-await.html:55)
    1) doTestChainedPromisesJSON (async-callstack-async-await.html:119)
    [Promise.resolve]
    0) doTestChainedPromisesJSON (async-callstack-async-await.html:118)
    [Promise.resolve]
    0) doTestChainedPromisesJSON (async-callstack-async-await.html:117)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestChainedPromisesJSON (async-callstack-async-await.html:116)
    3) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)

Call stack:
    0) thenCallback (async-callstack-async-await.html:55)
    1) doTestPromiseAll (async-callstack-async-await.html:129)
    [Promise.resolve]
    0) resolvePromise (async-callstack-async-await.html:12)
    [setTimeout]
    0) promiseCallback (async-callstack-async-await.html:21)
    1) timeoutPromise (async-callstack-async-await.html:9)
    2) doTestPromiseAll (async-callstack-async-await.html:128)
    3) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) thenCallback (async-callstack-async-await.html:55)
    1) doTestPromiseConstructor (async-callstack-async-await.html:70)
    [Promise.resolve]
    0) doTestPromiseConstructor (async-callstack-async-await.html:66)
    1) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) thenCallback (async-callstack-async-await.html:55)
    1) doTestPromiseResolve (async-callstack-async-await.html:179)
    [Promise.resolve]
    0) doTestPromiseResolve (async-callstack-async-await.html:178)
    1) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) thenCallback (async-callstack-async-await.html:55)
    1) doTestSettledPromisesResolved (async-callstack-async-await.html:80)
    [Promise.resolve]
    0) doTestSettledPromisesResolved (async-callstack-async-await.html:79)
    1) testFunctionTimeout (async-callstack-async-await.html:50)
    [setTimeout]
    0) testFunction (async-callstack-async-await.html:43)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
